% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-family.R
\name{fit.family}
\alias{fit.family}
\title{Fit family}
\usage{
fit.family(
  x,
  w = rep(1, length(x)),
  family = c("huber", "trimmed"),
  scale.fun = stats::mad,
  eps = .Machine$double.eps
)
}
\arguments{
\item{x}{a numeric vector of data}

\item{w}{a numeric vector of weights}

\item{family}{the location family; currently allows 'huber' for Huber family (default) or
'trimmed' for trimmed mean family}

\item{scale.fun}{a function used to estimate the scale \code{x} for Huber loss; ensures that the
tuning parameter is comparable across variables with different scales}

\item{eps}{a numerical tolerance parameter}
}
\value{
An object of class \code{fit.family}; a data frame with the following columns:
\item{mu.hat}{the fitted values}
\item{lambda}{the indexing parameter}
}
\description{
Fits the Huber or trimmed mean location families.
}
\examples{
fit <- fit.family(MASS::galaxies)
plot(fit)
}
\author{
Ryan Thompson <ryan.thompson@monash.edu>
}
