% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-family.R
\name{plot.fit.family}
\alias{plot.fit.family}
\title{Plot function for \code{fit.family} object}
\usage{
\method{plot}{fit.family}(x, y = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{fit.family}}

\item{y}{an object of class \code{fit.family}}

\item{...}{any other arguments}
}
\value{
A plot of the fitted family.
}
\description{
Plot a fitted family.
}
\author{
Ryan Thompson <ryan.thompson@monash.edu>
}
