% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center-test.R
\name{print.center.test}
\alias{print.center.test}
\title{Print function for \code{center.test} object}
\usage{
\method{print}{center.test}(x, ...)
}
\arguments{
\item{x}{an object of class \code{center.test}}

\item{...}{any other arguments}
}
\value{
The argument \code{x}.
}
\description{
Print objects of class \code{center.test}.
}
\author{
Ryan Thompson <ryan.thompson-1@uts.edu.au>
}
