% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet.R
\name{rudirichlet}
\alias{rudirichlet}
\title{Uniform Dirichlet distribution}
\usage{
rudirichlet(n, d)
}
\arguments{
\item{n}{the number of observations}

\item{d}{the number of dimensions}
}
\value{
A matrix; each row is a random draw and each column is a dimension.
}
\description{
Random number generation for the uniform Dirichlet distribution (having all
concentration parameters set to one).
}
\author{
Ryan Thompson <ryan.thompson1@unsw.edu.au>
}
