% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted-stats.R
\name{weighted}
\alias{weighted}
\alias{weighted.median}
\alias{weighted.mad}
\title{Weighted statistics}
\usage{
weighted.median(x, w)

weighted.mad(x, w)
}
\arguments{
\item{x}{a numeric vector of data}

\item{w}{a numeric vector of weights}
}
\value{
A length-one numeric vector.
}
\description{
Assorted weighted statistics unavailable in base R
}
\author{
Ryan Thompson <ryan.thompson@monash.edu>
}
