% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center-test.R
\name{plot.center.test}
\alias{plot.center.test}
\title{Plot function for \code{center.test} object}
\usage{
\method{plot}{center.test}(x, band = c(0.5, 0.75, 0.95), ...)
}
\arguments{
\item{x}{an object of class \code{center.test}}

\item{band}{a vector of band limits for functional box plot}

\item{...}{any other arguments}
}
\value{
A plot of the posterior distribution.
}
\description{
Plot the posterior distribution for the family of centers using a functional box plot.
}
\author{
Ryan Thompson <ryan.thompson@monash.edu>
}
